Edited Explodeable to remove objects on collision Edited Enemy behavior to no longer check for collisions Edited GameManager to change scoreMultiplier to a float

+++ Assets/Scripts/GameManager.cs
@@ -19,7 +19,7 @@ public class GameManager : MonoBehaviour {
 
     public int score;
 
-    public int scoreMultiplier;
+    public float scoreMultiplier;
 
     public GameObject player;
